telegram_bot/requirements.txt
python-telegram-bot
telegram
python-dotenv
sqlalchemy
Flask
requests